backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.114",
    "uvicorn[standard]>=0.32",
    "msal>=1.30",
    "httpx>=0.27",
    "pydantic>=2.6",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "PyYAML>=6.0",
    "cryptography>=43.0",
]
